三个模块请求方式改变以及沟通模块增加功能

+++ service/src/main/java/com/scrm/service/dao/CommunicationDao.java
@@ -18,7 +18,7 @@ public interface CommunicationDao {
 
     Integer editCommunication(Communication communication);
 
-    Integer deleteCommunication(Integer id);
+    Integer deleteCommunication(Long id);
 
-    Integer deleteCommunicationLog(Integer id);
+    Integer deleteCommunicationLog(Long id);
 }
